More Mapping
More mapping, saved multiple plots in the Plots directory. The title of
the plot pdf tells you the gamma setting used to generate that plot, and
the title tells you the alpha value used to generate that plot.

Also cleaned up the code a lot, moved the barnes analysis to its own
file which will have other analysis tools as well in the future.